testing with postgresql on Travis (#131)
* trying...

* fixing error

* trying to destroy db before terminating app

* without terminate?

* terminate with clear db

* fixing small error

* fixed possible async issue

* working locally, checking with postgres on travis

* trying to drop database

* oops. wrong database name

* have travis drop the database before creating it

* fixed travis config

* drop after destroy?

* drop or destroy. not both?

* modify migration files?

* ????

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* fixed model tests?

* added document to publication that exists

* activity now belongs to reader that exists

* still fixing foreign key contraints

+++ tests/integration/document.test.js
@@ -7,10 +7,11 @@ const {
   destroyDB,
   getActivityFromUrl
 } = require('./utils')
-const app = require('../../server').app
 
-const test = async () => {
-  await app.initialize()
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
 
   const token = getToken()
   const userId = await createUser(app, token)
@@ -79,8 +80,10 @@ const test = async () => {
     await tap.equal(res.statusCode, 404)
   })
 
-  await app.terminate()
-  await destroyDB()
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
 }
 
-test()
+module.exports = test
