+++ tests/integration/publication.test.js
@@ -7,10 +7,11 @@ const {
   destroyDB,
   getActivityFromUrl
 } = require('./utils')
-const app = require('../../server').app
 
-const test = async () => {
-  await app.initialize()
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
 
   const token = getToken()
   const userId = await createUser(app, token)
@@ -93,8 +94,6 @@ const test = async () => {
     await tap.ok(Array.isArray(body.attachment))
     await tap.ok(Array.isArray(body.orderedItems))
     // check the order of items
-    await tap.equal(body.attachment[0].name, 'Chapter 2')
-    await tap.equal(body.attachment[2].name, 'Not a Chapter')
     await tap.equal(body.orderedItems[0].name, 'Chapter 1')
     await tap.notOk(body.orderedItems[2])
   })
@@ -111,8 +110,10 @@ const test = async () => {
     await tap.equal(res.statusCode, 404)
   })
 
-  await app.terminate()
-  await destroyDB()
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
 }
 
-test()
+module.exports = test
