+++ tests/integration/reader.test.js
@@ -2,10 +2,11 @@ const request = require('supertest')
 const tap = require('tap')
 const urlparse = require('url').parse
 const { getToken, destroyDB } = require('./utils')
-const app = require('../../server').app
 
-const test = async () => {
-  await app.initialize()
+const test = async app => {
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.initialize()
+  }
 
   const token = getToken()
   let readerUrl
@@ -89,8 +90,10 @@ const test = async () => {
     await tap.equal(res.statusCode, 404)
   })
 
-  await app.terminate()
-  await destroyDB()
+  if (!process.env.POSTGRE_INSTANCE) {
+    await app.terminate()
+  }
+  await destroyDB(app)
 }
 
-test()
+module.exports = test
